Comment out dead and currently unworkable code
PLEASE ENABLE "ISSUES" FOR THIS REPOSITORY IT IS IMPOSSIBLE TO DISCUSS
THINGS LIKE THIS WITHOUT THEM.

I am guessing that you intended to include the (left shifted by 8 bits)
status register in the return value from `GetQuery(...)` but never got
around to do it. As such the commissioning code is not going to be
reliable IMHO - and this bug has been in since 8f6d9fbfc0d90c99718ace983109f6e61aa9893c.

I've raised this PR as a means to alert you that the line in the top level
README.md:
> Pull requests are welcome. For major changes, please open an issue first
to discuss what you would like to change.

is impossible if you haven't enabled "issues"!

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>